% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hello.R
\name{hello}
\alias{hello}
\title{Hello World}
\usage{
hello(who = "fuckwaddle", lang = "en", langdata = language)
}
\arguments{
\item{who}{name (string) of recipient}

\item{lang}{language code (string) : available : en it fr es de}

\item{langdata}{description}
}
\value{
\code{'Hello, world!'}
}
\description{
\code{hello} says \emph{'Hello, world!'}
}
