% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhello-package.R
\docType{package}
\name{myhello-package}
\alias{myhello}
\alias{myhello-package}
\title{myhello: Say hello in different languages!}
\description{
The package enables us to say hello in different languages!
}
\author{
\strong{Maintainer}: Léa Jacquet \email{lea.jacquet@student.uclouvain} (cre)

}
\keyword{internal}
