% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhello.R
\name{myhello}
\alias{myhello}
\title{Hello World}
\usage{
myhello(who, lang, LangData)
}
\arguments{
\item{who}{a \code{character} vector of length 1 that specifies the name of the person to whom the message is addressed.}

\item{lang}{a \code{character} vector of length 1 that specifies the preferred language. Default to "EN" for English. Other possible values include "FR" for French, "IT"  for Italian, "ES" for Spanish, or "DE" for German. Case is ignored.}

\item{LangData}{an optional data.frame with two columns each of mode \code{character}. The first column gives the language codes and the second column gives the corresponding "hello" word. Default to \code{language}.
see \code{?language}}
}
\value{
\code{"Hello, world!"}

a \code{caracter} vector with a personalized \emph{"hello"} message.
}
\description{
\code{myhello} says \emph{"Hello"} in the user-specified language. The user is asked to give her/his name so that the hello message gets personalized.
}
